avoid slow-down when lots of empty <match/>'es in a rule
On long tag-sequences, there seemed to be a lot of unnecessary copying
of state-vectors going on (oldvector=newvector).

There's still copying of new_state going on earlier in
processME/process, but this at least takes care of the worst offender.
Should probably switch to having alive_states be a pointer …

Needs regression testing!

git-svn-id: https://svn.code.sf.net/p/apertium/svn/trunk/apertium-lex-tools@73987 72bbbca6-d526-0410-a7d9-f06f51895060