support macros (#96)
Closes #11
Closes #33
Closes #55

This PR adds support for macros. As with previous metalrx features, this change will have no effect until references to `apertium-metalrx` are removed from the makefile, so no immediate action will be needed on the part of pair maintainers if this is merged.

Macros are defined like this:

```xml
<def-macros>
  <def-macro n="some_name" npar="1" nodes="2">
    <rule> ... </rule>
    ...
  </def-macro>
```
Where `npar="1"` specifies that is macro takes 1 string parameter and `nodes="2"` that it takes 2 node parameters.

The macro is then invoked as

```xml
<rules>
  <macro n="some_name">
    <with-param v="1st string param"/>
    <with-param>
      <match lemma="..."/>
    </with-param>
    <with-param>
      <match/>
      <match tags="sent"/>
    </with-param>
  </macro>
</rules>
```
That is, `<with-param>` containing other nodes is a node parameter and `<with-param>` with the `v=` attribute is a string parameter.

Node parameters can be inserted into a rule with `<param n="1"/>` and string parameters by placing a `p` before an attribute like `<match plemma="1"/>`.